% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DebtClasses.R
\docType{class}
\name{Debt-classes}
\alias{Debt-classes}
\alias{LogitDebt-class}
\title{S4 classes to model the effect of debt on mergers}
\description{
Extend classes from the \pkg{antitrust} package to accommodate horizontal mergers with debt.

The \dQuote{LogitDebt} class has the information for a Nash Bertand Pricing Game with Logit demand and firm-level debt. Note that code supports firms operating in multiple markets.

Let k denote the number of products produced by all firms across all markets. Let f denote the number of firms producing products across all markets. Let m denote the number of markets.
}
\section{Slots}{

\describe{
\item{\code{marketID}}{A length k character vector identifying which market a product is sold in.}

\item{\code{productID}}{A length k character vector identifying a product.}

\item{\code{debtPre}}{A length f vector of pre-merger debt payments.}

\item{\code{debtPost}}{A length f vector of post-merger debt payments. Default is \sQuote{debtPre}.}

\item{\code{density}}{A length 1 character vector equal to either "beta' (default) or "uniform". If "beta" calibration routine assumes that firm outside shares are sampled from a \code{\link[base]{Beta}} distribution and uses margin information to calibrate \sQuote{shape1} and \sQuote{shape2}. If "uniform", then \sQuote{shape1}=\sQuote{shape2}=1.}

\item{\code{focal}}{An integer (default 1) indicating which market is the "focal" market.}

\item{\code{parmsStart}}{A length $m + 2$ vector whose first $m$ elements are starting values for the price coefficients for each of the m markets. The final 2 elements are starting values for \sQuote{shape1} and \sQuote{shape2} of the \code{\link[base]{Beta}} distribution.}

\item{\code{shareOutParm}}{TBD}

\item{\code{hfocalIntegral}}{TBD}

\item{\code{gfocalIntegral}}{TBD}

\item{\code{tOtherIntegral}}{TBD}

\item{\code{rOtherIntegral}}{TBD}

\item{\code{control.cub}}{TBD}
}}

\section{Objects from the Class}{

For LogitDebt, objects can be created by using the constructor function \code{\link{logit.debt}}.
}

\section{Extends}{

LogitDebt: Class \code{\linkS4class{Logit}}, directly.
Class \code{\linkS4class{Bertran}}, by class \code{\linkS4class{Logit}}, distance 2.
}

\examples{
showClass("LogitDebt")           # get a detailed description of the class
}
\author{
Charles Taragin \email{ctaragin+mergerdebtr@gmail.com}
}
