% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DebtFunctions.R
\name{Debt-functions}
\alias{Debt-functions}
\alias{debt}
\alias{logit.debt}
\title{Merger Simulation with Debt}
\usage{
logit.debt(
  prices,
  shares,
  margins,
  ownerPre,
  ownerPost,
  debtPre = rep(0, length(prices)),
  debtPost = debtPre,
  marketID = as.character(rep(1, length(prices))),
  productID,
  mktSize = rep(1, unique(marketID)),
  density = c("beta", "uniform"),
  normIndex = 1,
  focal = 1,
  mcDelta = rep(0, length(prices)),
  priceOutside = 0,
  priceStart = prices,
  parmsStart,
  shareOutParm,
  isMax = FALSE,
  control.slopes,
  control.equ,
  control.cub,
  labels,
  ...
)
}
\arguments{
\item{prices}{A length k vector of  product prices.}

\item{shares}{A length k vector of product (quantity) shares. Values must be
between 0 and 1 .}

\item{margins}{A length k vector of product margins in proportions (bounded between 0 and 1), some of which may
equal NA.}

\item{ownerPre}{A length k vector whose values
indicate which firm produced a product pre-merger.}

\item{ownerPost}{A length k vector whose values
indicate which firm produced a product post-merger.}

\item{debtPre}{a length k vector of pre-merger FIRM-level debt payments. Note that sum of product-level debt payments should equal firm level debt.}

\item{debtPost}{a length k vector of post-merger debt payments. Default equals \sQuote{debtPre}.}

\item{marketID}{A length k vector whose values
indicate which market a product is sold in.}

\item{productID}{A length k vector whose values
indicate a product's identity.}

\item{mktSize}{A length m vector equal to market size.
If shares sum to one, this also equals the size of the market.}

\item{mcDelta}{A length k vector where each element equals the
proportional change in a downstream firm's product-level marginal costs due to
  the merger. Default is 0, which assumes that the merger does not
  affect any products' marginal cost.}

\item{priceOutside}{A length 1 vector indicating the price of the
outside good. Default is 0.}

\item{priceStart}{A length k vector of  starting values used to solve for
downstream equilibrium prices. Default is the \sQuote{prices} matrix.}

\item{control.slopes}{A list of  \code{\link{optim}}
control parameters passed to the calibration routine
optimizer (typically the \code{calcSlopes} method).}

\item{control.equ}{A list of  \code{\link[BB]{BBsolve}} control parameters
passed to the non-linear equation solver (typically the \code{calcPrices} method).}

\item{labels}{A k-length vector of labels. Default is "Prod#", where
\sQuote{#} is a number between 1 and the length of \sQuote{prices}.}

\item{...}{Additional options to feed to the \code{\link[BB]{BBsolve}}
optimizer used to solve for equilibrium prices.}
}
\value{
\code{logit.debt} returns an instance of class
\code{\linkS4class{LogitDebt}}.
}
\description{
Calibrates consumer demand using Logit
and then simulates the price effect of a merger between two firms
under the assumption that all firms in the market are playing a
differentiated products Bertrand pricing game with firm-level debt.

Let k denote the number of products produced by all firms across all markets. Let f denote the number of firms producing products across all markets. Let m denote the number of markets.
}
\details{
Using product prices, quantity shares and all of the
product margins from at least one firm, \code{logit} is able to
recover the price coefficient and product mean valuations in a
Logit demand model. \code{logit} then uses these
calibrated parameters to simulate a merger between two firms.
}
\examples{
## 3 single-product firms, each selling in 2 markets.
## Firms 1 and 2 merge
nprods <- 3
nmarkets <- 2
marketId <- rep(1:nmarkets,nprods)
shares <- c(0.156, 0.418, 0.616, 0.424, 0.229, 0.158)
prices <- c(97.303, 119.083, 94.103, 147.905, 93.113, 100.935)
margins <-c(17.163, 84.026, 26.451, 103.05, 17.316, 64.105)/prices
ownerPre <- ownerPost <- as.character(rep(c(1, 2, 3),each=nmarkets))
ownerPost[ownerPost=='2'] <- '1'
debt <- c(22.69,NA, 51.002,NA, 2.922,NA)
mktSize <- 1


simres_debt <- logit.debt(prices,shares,margins,
                             ownerPre,ownerPost,debtPre=debt,marketID=marketId,
                             mktSize=mktSize)


print(simres_debt)
summary(simres_debt)

}
\author{
Charles Taragin \email{ctaragin+antitrustr@gmail.com}
}
