% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debt-methods.R
\name{LogitDebt-methods}
\alias{LogitDebt-methods}
\alias{calcShares,}
\alias{calcSlopes,}
\alias{calcPrices,}
\alias{calcMargins,}
\alias{calcMC,}
\alias{summary,}
\alias{plotBeta,LogitDebt-method}
\alias{calcShares,LogitDebt-method}
\alias{calcSlopes,LogitDebt-method}
\alias{calcPrices,LogitDebt-method}
\alias{calcMC,LogitDebt-method}
\alias{calcMargins,LogitDebt-method}
\alias{summary,LogitDebt-method}
\title{Additional methods for LogitDebt Class}
\usage{
\S4method{plotBeta}{LogitDebt}(object)

\S4method{calcShares}{LogitDebt}(
  object,
  preMerger = TRUE,
  revenue = FALSE,
  insideOnly = FALSE,
  outSideOnly = FALSE
)

\S4method{calcSlopes}{LogitDebt}(object)

\S4method{calcPrices}{LogitDebt}(object, preMerger = TRUE, isMax = FALSE, ...)

\S4method{calcMC}{LogitDebt}(object, preMerger = TRUE)

\S4method{calcMargins}{LogitDebt}(object, preMerger = TRUE, level = FALSE)

\S4method{summary}{LogitDebt}(
  object,
  revenue = FALSE,
  shares = TRUE,
  levels = FALSE,
  parameters = FALSE,
  market = FALSE,
  insideOnly = TRUE,
  digits = 2,
  ...
)
}
\arguments{
\item{object}{an instance of class \code{LogitDebt}}

\item{preMerger}{when TRUE, computes equilibrium prices under the pre-merger regime. When FALSE, calculates
post-merger equilibrium prices.Default is TRUE.}

\item{revenue}{when TRUE, computes revenue shares. When FALSE, calculates
quantity shares. Default is FALSE.}

\item{...}{harmlessly pass the arguments used in other calcPrices methods to
methods for \code{LogitDebt}.}

\item{level}{when TRUE, computes margins in dollars. When FALSE, calculates
margins as a proportion of prices. Default is FALSE.}
}
\value{
\code{calcSlopes} return a  \code{LogitDebt}  object containing estimated slopes. \code{CalcQuantities} returns
a matrix of equilibrium quantities under `preMerger'.
}
\description{
\code{calcShares}, \code{calcSlopes}, \code{calcPrices}, \code{calcMargins}, \code{calcMC},
\code{summary} methods for the \code{LogitDebt} class
}
